// src/infraestructure/models/PokemonInfoModel.ts
export interface PokemonInfoModel {
    abilities:                Ability[];
    base_experience:          number;
    cries:                    Cries;
    forms:                    Species[];
    height:                   number;
    id:                       number;
    name:                     string;
    order:                    number;
    species:                  Species;
    sprites:                  Sprites;
    weight:                   number;
}

export interface Ability {
    ability:   Species;
    is_hidden: boolean;
    slot:      number;
}

export interface Species {
    name: string;
    url:  string;
}

export interface Cries {
    latest: string;
    legacy: string;
}



export interface Sprites {
    back_default:       string;
    back_female:        null;
    back_shiny:         string;
    back_shiny_female:  null;
    front_default:      string;
    front_female:       null;
    front_shiny:        string;
    front_shiny_female: null;
    animated?:          Sprites;
}